+++ test/cpp/jit/test_backend.cpp
@@ -1,81 +1,40 @@
-#include <torch/csrc/jit/backends/backend.h>
+#include <gtest/gtest.h>
+#include <torch/csrc/jit/api/module.h>
+#include <torch/csrc/jit/backends/backend_detail.h>
+#include <torch/torch.h>
 
+// Tests go in torch::jit
 namespace torch {
 namespace jit {
-// This test JIT backend is intended to do the minimal amount of work
-// necessary to test that the JIT backend registration endpoints and
-// code generation are working correctly. It is not intended to
-// produce numerically correct results.
-class TestBackend : public PyTorchBackendInterface {
- public:
-  // Constructor.
-  explicit TestBackend() {}
-  virtual ~TestBackend() = default;
+TEST(BackendTest, ToBackend) {
+  Module m("m");
+  m.define(R"(
+    def forward(self, x, h):
+        return self.accum(x, h), self.sub_accum(x, h)
 
-  c10::IValue preprocess(
-      c10::IValue mod,
-      c10::impl::GenericDict method_compile_spec) override {
-    return mod;
-  }
+    def accum(self, x, h):
+        return x + h
 
-  c10::impl::GenericDict compile(
-      c10::IValue processed,
-      c10::impl::GenericDict method_compile_spec) override {
-    auto spec =
-        c10::impl::toTypedDict<std::string, at::IValue>(method_compile_spec);
+    def sub_accum(self, x, h):
+        return x - h
+  )");
 
-    // Return the same string as a value for every key in method_compile_spec.
-    auto handles = c10::Dict<std::string, std::string>();
-    for (const auto& it : spec) {
-      handles.insert(it.key(), it.key());
-    }
-    return c10::impl::toGenericDict(handles);
-  }
-  c10::impl::GenericList execute(
-      c10::IValue handle,
-      c10::impl::GenericList inputs) override {
-    TORCH_INTERNAL_ASSERT(handle.isString());
-    TORCH_INTERNAL_ASSERT(inputs.size() > 0);
+  std::vector<IValue> inputs;
+  inputs.emplace_back(2.0 * torch::ones({}));
+  inputs.emplace_back(1.0 * torch::ones({}));
+  auto ref = m.forward(inputs).toTuple()->elements();
 
-    c10::List<at::Tensor> output_list;
-
-    // Implement simple accumulator and negative accumulator (?) ops. Return one
-    // or both of them depending on the handle to make sure multiple outputs are
-    // handled.
-    c10::IValue value = inputs[0];
-    at::Tensor accum = value.toTensor();
-    accum = accum.clone();
-    at::Tensor sub_accum = value.toTensor();
-    sub_accum = sub_accum.clone();
-
-    for (size_t i = 1, e = inputs.size(); i < e; ++i) {
-      value = inputs[i];
-      accum.add_(value.toTensor(), 1.0);
-      sub_accum.sub_(value.toTensor(), 1.0);
-    }
-
-    if (handle.toStringRef() == "accum") {
-      output_list.emplace_back(accum);
-    } else if (handle.toStringRef() == "sub_accum") {
-      output_list.emplace_back(sub_accum);
-    } else if (handle.toStringRef() == "forward") {
-      output_list.emplace_back(accum);
-      output_list.emplace_back(sub_accum);
-    }
-
-    return c10::impl::toList(output_list);
-  }
-};
-
-c10::IValue preprocess(
-    const Module& mod,
-    const c10::Dict<IValue, IValue>& method_compile_spec) {
-  return mod._ivalue();
+  c10::Dict<IValue, IValue> compile_spec(StringType::get(), AnyType::get());
+  c10::Dict<IValue, IValue> fake_dict(StringType::get(), AnyType::get());
+  fake_dict.insert("", "");
+  compile_spec.insert("forward", fake_dict);
+  auto any_dict_ty = DictType::create(StringType::get(), AnyType::get());
+  // lowered module
+  auto lm = torch::jit::detail::codegen_backend_module(
+      "test_backend", m, compile_spec, any_dict_ty);
+  auto res = lm.forward(inputs).toTuple()->elements();
+  AT_ASSERT(res[0].toTensor().equal(ref[0].toTensor()));
+  AT_ASSERT(res[1].toTensor().equal(ref[1].toTensor()));
 }
-
-namespace {
-static auto cls = torch::jit::backend<TestBackend>("test_backend", preprocess);
-}
-
 } // namespace jit
 } // namespace torch
