+++ torch/csrc/jit/backends/backend_detail.h
@@ -58,11 +58,15 @@ TORCH_API void registerBackendPreprocessFunction(
     const std::string& name,
     const BackendPreprocessFunction& preprocess);
 
-TORCH_API bool hasBackendPreprocessFunction(const std::string& name);
+bool hasBackendPreprocessFunction(const std::string& name);
 
-TORCH_API BackendPreprocessFunction
-getBackendPreprocessFunction(const std::string& name);
+BackendPreprocessFunction getBackendPreprocessFunction(const std::string& name);
 
+TORCH_API Module codegen_backend_module(
+    const std::string& backend_name,
+    const Module& orig_module,
+    const c10::Dict<IValue, IValue>& method_compile_spec,
+    const c10::DictTypePtr& any_dict_ty);
 } // namespace detail
 } // namespace jit
 } // namespace torch
